fetch data on the navigation segment's toggle
